serveur/storage/requirements.txt
flask
psycopg2-binary
docling
docling-core
langchain-core
langchain-qdrant
langchain-ollama
litellm
pydantic
python-dotenv